+++ docs/changelog.rst
@@ -24,6 +24,11 @@ Changelog
 
   `PR #643 <https://github.com/adamchainz/time-machine/pull/643>`__.
 
+* Support isolated subinterpreters, as created by |concurrent.interpreters|__ on Python 3.14+, on Python 3.12+.
+
+  .. |concurrent.interpreters| replace:: ``concurrent.interpreters``
+  __ https://docs.python.org/3.14/library/concurrent.interpreters.html
+
 * Build with frame pointers enabled, preparation for `PEP 831 <https://peps.python.org/pep-0831/>`__.
 
   `PR #627 <https://github.com/adamchainz/time-machine/issues/627>`__.